third_party/WebKit/LayoutTests/external/wpt/safe-html-types/SafeHTML.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="./support/helper.js"></script>
<script>
  promise_test(t => {
    return SafeHTML.escape(unescapedHTML).then(html => {
      assert_equals(html.toString(), escapedHTML);
    });
  }, "Basic escaping.");

  promise_test(t => {
    return SafeHTML.unsafelyCreate(unescapedHTML).then(html => {
      assert_equals(html.toString(), unescapedHTML);
    });
  }, "Basic unsafe construction.");
</script>
